frontend/requirements.txt
streamlit
sseclient-py
plotly
pandas